+++ Changelog.md
@@ -4,10 +4,12 @@ Language Features:
 
 
 Compiler Features:
+ * SMTChecker: Support inherited state variables.
 
 
 Bugfixes:
  * SMTChecker: Fix bad cast in base constructor modifier.
+ * SMTChecker: Fix internal error when visiting state variable inherited from base class.